LPS-73056 Remove download command, we don't need it for now
